requirements.txt
Flask
mysql-connector-python
tmdbsimple
pandas
numpy
tensorflow
keras
PyYAML
